+++ eclair-core/src/main/scala/fr/acinq/eclair/blockchain/OnChainWallet.scala
@@ -90,7 +90,11 @@ trait OnChainAddressGenerator {
 }
 
 trait OnchainPubkeyCache {
-    def getP2wpkhPubkey(renew: Boolean = true): PublicKey
+
+  /**
+   * @param renew applies after requesting the current pubkey, and is asynchronous
+   */
+  def getP2wpkhPubkey(renew: Boolean = true): PublicKey
 }
 
 /** This trait lets users check the wallet's on-chain balance. */
